NGINX FastCGI
The FastCGI protocol is designed for communication between a web server and a web application. It is commonly used for PHP or Python web applications. However, when I needed it, there was no usable implementation for Delphi available. This library fills that gap. It is MIT licensed, tested and documented.